Renamed folders, interface/class names, and file names.
