+++ GermanVocabApp.Api.FluentValidation/VocabListValidationController.cs
@@ -1,23 +1,23 @@
 ﻿using FluentValidation.Results;
 using GermanVocabApp.Api.FluentValidation;
-using GermanVocabApp.Api.FluentValidation.FluentValidators;
+using GermanVocabApp.Api.FluentValidation.Validators;
 using GermanVocabApp.Core.Contracts;
 
 namespace GermanVocabApp.Api.VocabLists.Validation;
 
-public class VocabListRequestValidator<TItem> : IValidator<IListRequest<TItem>>
+public class VocabListValidationController<TItem> : IValidationController<IListRequest<TItem>>
     where TItem : IListItemRequest
 {
     private readonly FluentListValidator<TItem> _listValidator;
     private readonly WordValidatorFactory _wordValidatorFactory;
 
-    public VocabListRequestValidator()
+    public VocabListValidationController()
     {
         _listValidator = new FluentListValidator<TItem>();
         _wordValidatorFactory = new WordValidatorFactory();
     }
 
-    internal VocabListRequestValidator(FluentListValidator<TItem> listValidator, WordValidatorFactory wordValidatorFactory)
+    internal VocabListValidationController(FluentListValidator<TItem> listValidator, WordValidatorFactory wordValidatorFactory)
     {
         _listValidator = listValidator;
         _wordValidatorFactory = wordValidatorFactory;
